# openapi.yaml
openapi: 3.1.0
info:
  title: Agent Service Interop Inspection API
  version: 0.3.0-experimental
  description: >-
    Read-only projections over the same underlying service workflow exposed
    through the repository's A2A inspection skill. These APIs do not authorize,
    schedule, complete, settle, or otherwise mutate a real-world service job.
paths:
  /api/interop/workflows/{workflow_id}/inspection:
    get:
      operationId: inspectServiceWorkflow
      summary: Inspect an existing service workflow
      description: >-
        Local/research read-only interoperability observation for an existing
        workflow. This route is not the public x402 stranger-test surface.
      parameters:
        - name: workflow_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^wf-.+$'
      responses:
        '200':
          description: Read-only workflow inspection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowInspection'
        '404':
          description: Workflow not found
  /api/x402/public/workflow-inspection:
    get:
      operationId: inspectPublicSyntheticWorkflow
      summary: Inspect the configured public synthetic workflow
      description: >-
        Paid x402 resource for one server-configured synthetic/public workflow.
        The caller does not supply a workflow identifier. Payment authorizes only
        this digital resource invocation; it does not establish identity,
        delegation, access to private workflows, quote acceptance, service
        authorization, service payment, fulfillment, or customer acceptance.
      responses:
        '200':
          description: Paid read-only inspection of the configured public synthetic workflow
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowInspection'
        '402':
          description: x402 payment required
        '404':
          description: Configured public synthetic resource is unavailable
components:
  schemas:
    WorkflowInspection:
      type: object
      required: [semantics, references, facets]
      properties:
        semantics:
          type: object
          required: [interaction_state_meaning]
          properties:
            interaction_state_meaning:
              type: string
          additionalProperties: false
        references:
          type: object
          required:
            - workflow_id
            - operational_refs
            - protocol_refs
          properties:
            workflow_id:
              type: string
              description: Repository-local protocol-neutral correlation identifier; not an operational authority identifier.
            operational_refs:
              type: array
              items:
                type: object
                required: [system, object_type, id]
                properties:
                  system:
                    type: string
                  object_type:
                    type: string
                  id:
                    type: string
                additionalProperties: false
            protocol_refs:
              type: array
              items:
                type: object
                required: [protocol, object_type, id]
                properties:
                  protocol:
                    type: string
                  object_type:
                    type: string
                  id:
                    type: string
                additionalProperties: false
          additionalProperties: false
        facets:
          type: object
          properties:
            quote:
              type: object
              required: [status]
              properties:
                status:
                  type: string
              additionalProperties: false
            job:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                scheduled_for:
                  type: [string, 'null']
                  format: date-time
              additionalProperties: false
          additionalProperties: false
      additionalProperties: false